apps/agent-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Event-driven agent orchestration service"
requires-python = ">=3.10"
dependencies = [
    "aio-pika>=9.0",
    "httpx>=0.25.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0",
    "azure-ai-projects>=2.0.0b1",
    "azure-identity>=1.15.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
agent-orchestrator = "src.orchestrator:main"
